requirements.txt
openpyxl==3.1.5
requests==2.32.3
ttkbootstrap==1.12.0
python-dateutil==2.9.0.post0

# Optional: drag-and-drop support for Tkinter. App will gracefully fall back if missing.
tkinterdnd2==0.4.3

# Dev/test
pytest==8.3.4

